Remove ConsoleGems dependency on Microsoft.Extensions.DependencyInjection
